scripts/python/requirements.txt
numpy
pandas
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib